+++ Unity_Project/Assets/Design.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ce7064ce4fdada24c88df4f343c65d87
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
