+++ Unity_Project/Assets/Program/Scenes/MousePoint.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bfb378e5e4674e34dad3f395f4ffde83
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
