--- Unity_Project/Assets/Scenes.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7b27825e8b14c174284a1c5713ceae2f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
