+++ Unity_Project/Library/Bee/artifacts/1900b0aE.dag/Assembly-CSharp.rsp
@@ -356,7 +356,7 @@
 -r:"Library/Bee/artifacts/1900b0aE.dag/Unity.VSCode.Editor.ref.dll"
 -r:"Library/Bee/artifacts/1900b0aE.dag/UnityEditor.UI.ref.dll"
 -r:"Library/Bee/artifacts/1900b0aE.dag/UnityEngine.UI.ref.dll"
-"Assets/MousePoint.cs"
+"Assets/Program/Scenes/MousePoint.cs"
 -langversion:9.0
 
 /deterministic
