+++ Unity_Project/Library/LastSceneManagerSetup.txt
@@ -1,5 +1,5 @@
 sceneSetups:
-- path: Assets/Scenes/SampleScene.unity
+- path: Assets/Program/Scenes/SampleScene.unity
   isLoaded: 1
   isActive: 1
   isSubScene: 0
