+++ Unity_Project/Library/StateCache/SceneView/67/6732087386cc0d711826836a5ec2e82c.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"m_AlwaysRefresh":false,"showFog":true,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":true,"pivot":{"x":746.2784423828125,"y":285.8229675292969,"z":1.775108814239502},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":449.19464111328127,"orthographic":true}
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"m_AlwaysRefresh":false,"showFog":true,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":true,"pivot":{"x":611.8659057617188,"y":297.15423583984377,"z":3.3349297046661379},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":293.2127685546875,"orthographic":true}
